reorganized code into seprate modules
