+++ pkg/interlace/builder.go
@@ -17,24 +17,21 @@
 package interlace
 
 import (
-	"bufio"
 	"fmt"
-	"io/ioutil"
-	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
+	"github.com/IBM/argocd-interlace/pkg/application"
 	"github.com/IBM/argocd-interlace/pkg/config"
 	"github.com/IBM/argocd-interlace/pkg/manifest"
-	"github.com/IBM/argocd-interlace/pkg/provenance"
+	helmprov "github.com/IBM/argocd-interlace/pkg/provenance/helm"
+	"github.com/IBM/argocd-interlace/pkg/provenance/kustomize"
+	kustprov "github.com/IBM/argocd-interlace/pkg/provenance/kustomize"
 	"github.com/IBM/argocd-interlace/pkg/storage"
 	"github.com/IBM/argocd-interlace/pkg/storage/annotation"
 	"github.com/IBM/argocd-interlace/pkg/utils"
 	appv1 "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 	log "github.com/sirupsen/logrus"
-	"golang.org/x/crypto/openpgp"
-	"golang.org/x/crypto/openpgp/packet"
 )
 
 func CreateEventHandler(app *appv1.Application) error {
@@ -47,26 +44,41 @@ func CreateEventHandler(app *appv1.Application) error {
 	appSourceRevision := app.Spec.Source.TargetRevision
 	appSourceCommitSha := ""
 	// Create does not have app.Status.Sync.Revision information, we need to extract commitsha by API
-	commitSha := provenance.GitLatestCommitSha(app.Spec.Source.RepoURL, app.Spec.Source.TargetRevision)
+	commitSha := kustomize.GitLatestCommitSha(app.Spec.Source.RepoURL, app.Spec.Source.TargetRevision)
 	if commitSha != "" {
 		appSourceCommitSha = commitSha
 	}
 
 	log.Infof("[INFO][%s]: Interlace detected creation of new Application resource: %s", appName, appName)
+	appPath := ""
+	isHelm := app.Spec.Source.IsHelm()
+	if isHelm {
+		appPath = fmt.Sprintf("%s/%s", "/tmp", appName)
+	} else {
+		appPath = app.Spec.Source.Path
+	}
 
-	appPath := app.Spec.Source.Path
 	appSourcePreiviousCommitSha := ""
 	var err error
 	sourceVerified := false
-	isHelm := app.Spec.Source.IsHelm()
+
+	appDirPath := filepath.Join(utils.TMP_DIR, appName, appPath)
+	chart := app.Spec.Source.Chart
+	appData, _ := application.NewApplicationData(appName, appPath, appDirPath, appClusterUrl,
+		appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha,
+		chart, isHelm)
 
 	if isHelm {
-		chart := app.Spec.Source.Chart
 		log.Infof("[INFO][%s]: Interlace detected creation of new Application resource: %s", appName, appName)
-		path := fmt.Sprintf("%s/%s", "/tmp", appName)
-		sourceVerified = verifyHelm(path, appSourceRepoUrl, chart, appSourceRevision)
+		prov, _ := helmprov.NewProvenance(*appData)
+		sourceVerified, err = prov.VerifySourceMaterial()
+		if err != nil {
+			log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials failed: %s", appName, appName)
+			return err
+		}
 	} else {
-		sourceVerified, err = verifySourceMaterial(appPath, appSourceRepoUrl)
+		prov, _ := kustprov.NewProvenance(*appData)
+		sourceVerified, err = prov.VerifySourceMaterial()
 
 		if err != nil {
 			log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials failed: %s", appName, appName)
@@ -77,9 +89,8 @@ func CreateEventHandler(app *appv1.Application) error {
 	if sourceVerified {
 		log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials succeeded: %s", appName, appName)
 
-		err = signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
-			appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha, true, isHelm,
-		)
+		err = signManifestAndGenerateProvenance(*appData, true)
+
 		if err != nil {
 			return err
 		}
@@ -132,12 +143,29 @@ func UpdateEventHandler(oldApp, newApp *appv1.Application) error {
 		log.Infof("[INFO][%s]: Interlace detected update of existing Application resource: %s", appName, appName)
 		isHelm := newApp.Spec.Source.IsHelm()
 		if isHelm {
-			chart := newApp.Spec.Source.Chart
-			log.Infof("[INFO][%s]: Interlace detected creation of new Application resource: %s", appName, appName)
-			path := fmt.Sprintf("%s/%s", "/tmp", appName)
-			sourceVerified = verifyHelm(path, appSourceRepoUrl, chart, appSourceRevision)
+			appPath = fmt.Sprintf("%s/%s", "/tmp", appName)
 		} else {
-			sourceVerified, err = verifySourceMaterial(appPath, appSourceRepoUrl)
+			appPath = newApp.Spec.Source.Path
+		}
+
+		appDirPath := filepath.Join(utils.TMP_DIR, appName, appPath)
+		chart := newApp.Spec.Source.Chart
+		appData, _ := application.NewApplicationData(appName, appPath, appDirPath, appClusterUrl,
+			appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha,
+			chart, isHelm)
+
+		if isHelm {
+
+			log.Infof("[INFO][%s]: Interlace detected creation of new Application resource: %s", appName, appName)
+			prov, _ := helmprov.NewProvenance(*appData)
+			sourceVerified, err = prov.VerifySourceMaterial()
+			if err != nil {
+				log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials failed: %s", appName, appName)
+				return err
+			}
+		} else {
+			prov, _ := kustprov.NewProvenance(*appData)
+			sourceVerified, err = prov.VerifySourceMaterial()
 
 			if err != nil {
 				log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials failed: %s", appName, appName)
@@ -149,8 +177,7 @@ func UpdateEventHandler(oldApp, newApp *appv1.Application) error {
 		if sourceVerified {
 			log.Infof("[INFO][%s]: Interlace's signature verification of Application source materials succeeded: %s", appName, appName)
 
-			err := signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
-				appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha, created, isHelm)
+			err := signManifestAndGenerateProvenance(*appData, created)
 			if err != nil {
 				return err
 			}
@@ -160,224 +187,7 @@ func UpdateEventHandler(oldApp, newApp *appv1.Application) error {
 	return nil
 }
 
-func verifyHelm(appPath, repoUrl, chart, targetRevision string) bool {
-	mkDirCmd := "mkdir"
-	_, err := utils.CmdExec(mkDirCmd, "", appPath)
-	helmChartUrl := fmt.Sprintf("%s/%s-%s.tgz", repoUrl, chart, targetRevision)
-	log.Info("[INFO]: appPath: ", appPath)
-	output := fmt.Sprintf("%s/%s-%s.tgz", appPath, chart, targetRevision)
-	curlCmd := "curl"
-	_, err = utils.CmdExec(curlCmd, appPath, helmChartUrl, "--output", output)
-	if err != nil {
-		log.Infof("[INFO]: Curl Helm Chart CmdExec download : %s ", err.Error())
-		return false
-	}
-
-	helmChartProvUrl := fmt.Sprintf("%s/%s-%s.tgz.prov", repoUrl, chart, targetRevision)
-	output = fmt.Sprintf("%s/%s-%s.tgz.prov", appPath, chart, targetRevision)
-	_, err = utils.CmdExec(curlCmd, appPath, helmChartProvUrl, "--output", output)
-	if err != nil {
-		log.Infof("[INFO]: Curl Helm Chart Prov CmdExec download : %s ", err.Error())
-		return false
-	}
-
-	helmCmd := "helm"
-	output = fmt.Sprintf("%s/%s-%s.tgz", appPath, chart, targetRevision)
-	_, err = utils.CmdExec(helmCmd, appPath, "sigstore", "verify", output)
-	if err != nil {
-		log.Infof("[INFO]: Helm Sigstore verify CmdExec add : %s ", err.Error())
-		return false
-	}
-
-	log.Info("[INFO]: Helm Sigstore verify was successfull")
-
-	return true
-	/*
-		interlaceConfig, err := config.GetInterlaceConfig()
-
-		hashFileUrl := fmt.Sprintf("%s/%s", repoUrl, interlaceConfig.SourceMaterialHashList)
-		output = fmt.Sprintf("%s/%s", appPath, interlaceConfig.SourceMaterialHashList)
-		_, err = utils.CmdExec(curlCmd, appPath, hashFileUrl, "--output", output)
-		if err != nil {
-			log.Infof("[INFO]: Curl Helm Chart Sourcematrial CmdExec download : %s ", err.Error())
-		}
-
-
-			signFileUrl := fmt.Sprintf("%s/%s", repoUrl, interlaceConfig.SourceMaterialSignature)
-			output = fmt.Sprintf("%s/%s", appPath, interlaceConfig.SourceMaterialSignature)
-			_, err = utils.CmdExec(curlCmd, appPath, signFileUrl, "--output", output)
-			if err != nil {
-				log.Infof("[INFO]: Curl Helm Chart Sourcematrial CmdExec download : %s ", err.Error())
-			}
-
-			keyPath := utils.KEYRING_PUB_KEY_PATH
-
-			srcMatPath := filepath.Join(appPath, interlaceConfig.SourceMaterialHashList)
-			srcMatSigPath := filepath.Join(appPath, interlaceConfig.SourceMaterialSignature)
-
-			verification_target, err := os.Open(srcMatPath)
-			signature, err := os.Open(srcMatSigPath)
-			flag, _, _, _, _ := verifySignature(keyPath, verification_target, signature)
-
-			log.Info("[INFO]: Helm source material verify was successfull ", flag)
-
-			return flag
-	*/
-}
-
-func verifySourceMaterial(appPath, appSourceRepoUrl string) (bool, error) {
-
-	interlaceConfig, err := config.GetInterlaceConfig()
-
-	host, orgRepo, path, gitRef, gitSuff := provenance.ParseGitUrl(appSourceRepoUrl)
-
-	log.Info("appSourceRepoUrl ", appSourceRepoUrl)
-	log.Info("host:", host, " orgRepo:", orgRepo, " path:", path, " gitRef:", gitRef, " gitSuff:", gitSuff)
-
-	url := host + orgRepo + gitSuff
-
-	log.Info("url:", url)
-
-	r, err := provenance.GetTopGitRepo(url)
-	if err != nil {
-		log.Errorf("Error git clone:  %s", err.Error())
-		return false, err
-	}
-
-	baseDir := filepath.Join(r.RootDir, appPath)
-
-	keyPath := utils.KEYRING_PUB_KEY_PATH
-
-	srcMatPath := filepath.Join(baseDir, interlaceConfig.SourceMaterialHashList)
-	srcMatSigPath := filepath.Join(baseDir, interlaceConfig.SourceMaterialSignature)
-
-	verification_target, err := os.Open(srcMatPath)
-	signature, err := os.Open(srcMatSigPath)
-	flag, _, _, _, _ := verifySignature(keyPath, verification_target, signature)
-
-	hashCompareSuccess := false
-	if flag {
-		hashCompareSuccess, err = compareHash(srcMatPath, baseDir)
-		if err != nil {
-			return hashCompareSuccess, err
-		}
-		return hashCompareSuccess, nil
-	}
-
-	return flag, nil
-}
-
-func verifySignature(keyPath string, msg, sig *os.File) (bool, string, *Signer, []byte, error) {
-
-	if keyRing, err := LoadKeyRing(keyPath); err != nil {
-		return false, "Error when loading key ring", nil, nil, err
-	} else if signer, err := openpgp.CheckArmoredDetachedSignature(keyRing, msg, sig); signer == nil {
-		if err != nil {
-			log.Error("Signature verification error:", err.Error())
-		}
-		return false, "Signed by unauthrized subject (signer is not in public key), or invalid format signature", nil, nil, nil
-	} else {
-		idt := GetFirstIdentity(signer)
-		fingerprint := ""
-		if signer.PrimaryKey != nil {
-			fingerprint = fmt.Sprintf("%X", signer.PrimaryKey.Fingerprint)
-		}
-		return true, "", NewSignerFromUserId(idt.UserId), []byte(fingerprint), nil
-	}
-}
-
-func GetFirstIdentity(signer *openpgp.Entity) *openpgp.Identity {
-	for _, idt := range signer.Identities {
-		return idt
-	}
-	return nil
-}
-
-type Signer struct {
-	Email              string `json:"email,omitempty"`
-	Name               string `json:"name,omitempty"`
-	Comment            string `json:"comment,omitempty"`
-	Uid                string `json:"uid,omitempty"`
-	Country            string `json:"country,omitempty"`
-	Organization       string `json:"organization,omitempty"`
-	OrganizationalUnit string `json:"organizationalUnit,omitempty"`
-	Locality           string `json:"locality,omitempty"`
-	Province           string `json:"province,omitempty"`
-	StreetAddress      string `json:"streetAddress,omitempty"`
-	PostalCode         string `json:"postalCode,omitempty"`
-	CommonName         string `json:"commonName,omitempty"`
-	SerialNumber       string `json:"serialNumber,omitempty"`
-	Fingerprint        []byte `json:"finerprint"`
-}
-
-func NewSignerFromUserId(uid *packet.UserId) *Signer {
-	return &Signer{
-		Email:   uid.Email,
-		Name:    uid.Name,
-		Comment: uid.Comment,
-	}
-}
-
-func LoadKeyRing(keyPath string) (openpgp.EntityList, error) {
-	entities := []*openpgp.Entity{}
-	var retErr error
-	kpath := filepath.Clean(keyPath)
-	if keyRingReader, err := os.Open(kpath); err != nil {
-		log.Warn("Failed to open keyring")
-		retErr = err
-	} else {
-		tmpList, err := openpgp.ReadKeyRing(keyRingReader)
-		if err != nil {
-			log.Warn("Failed to read keyring")
-			retErr = err
-		}
-		for _, tmp := range tmpList {
-			for _, id := range tmp.Identities {
-				log.Info("identity name ", id.Name, " id.UserId.Name: ", id.UserId.Name, " id.UserId.Email:", id.UserId.Email)
-			}
-			entities = append(entities, tmp)
-		}
-	}
-	return openpgp.EntityList(entities), retErr
-}
-
-func compareHash(sourceMaterialPath string, baseDir string) (bool, error) {
-	sourceMaterial, err := ioutil.ReadFile(sourceMaterialPath)
-
-	if err != nil {
-		log.Errorf("Error in reading sourceMaterialPath:  %s", err.Error())
-		return false, err
-	}
-
-	scanner := bufio.NewScanner(strings.NewReader(string(sourceMaterial)))
-
-	for scanner.Scan() {
-		l := scanner.Text()
-
-		data := strings.Split(l, " ")
-		if len(data) > 2 {
-			hash := data[0]
-			path := data[2]
-
-			absPath := filepath.Join(baseDir, "/", path)
-			computedFileHash, err := utils.ComputeHash(absPath)
-			log.Info("file: ", path, " hash:", hash, " absPath:", absPath, " computedFileHash: ", computedFileHash)
-			if err != nil {
-				return false, err
-			}
-
-			if hash != computedFileHash {
-				return false, nil
-			}
-		} else {
-			continue
-		}
-	}
-	return true, nil
-}
-
-func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
-	appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha string, created, isHelm bool) error {
+func signManifestAndGenerateProvenance(appData application.ApplicationData, created bool) error {
 
 	interlaceConfig, err := config.GetInterlaceConfig()
 	if err != nil {
@@ -387,11 +197,7 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 
 	manifestStorageType := interlaceConfig.ManifestStorageType
 
-	appDirPath := filepath.Join(utils.TMP_DIR, appName, appPath)
-
-	allStorageBackEnds, err := storage.InitializeStorageBackends(appName, appPath, appDirPath, appClusterUrl,
-		appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha, manifestStorageType, isHelm,
-	)
+	allStorageBackEnds, err := storage.InitializeStorageBackends(appData, manifestStorageType)
 
 	if err != nil {
 		log.Errorf("Error in initializing storage backends: %s", err.Error())
@@ -412,22 +218,22 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 
 		if created {
 			log.Info("created scenario")
-			log.Infof("[INFO][%s] Interlace downloads desired manifest from ArgoCD REST API", appName)
-			manifestGenerated, err = manifest.GenerateInitialManifest(appName, appPath, appDirPath)
+			log.Infof("[INFO][%s] Interlace downloads desired manifest from ArgoCD REST API", appData.AppName)
+			manifestGenerated, err = manifest.GenerateInitialManifest(appData)
 			if err != nil {
 				log.Errorf("Error in generating initial manifest: %s", err.Error())
 				return err
 			}
 		} else {
 			log.Info("update scenario")
-			log.Infof("[INFO][%s] Interlace downloads desired manifest from ArgoCD REST API", appName)
+			log.Infof("[INFO][%s] Interlace downloads desired manifest from ArgoCD REST API", appData.AppName)
 			yamlBytes, err := storageBackend.GetLatestManifestContent()
 			if err != nil {
 				log.Errorf("Error in retriving latest manifest content: %s", err.Error())
 
 				if storageBackend.Type() == annotation.StorageBackendAnnotation {
 					log.Info("Going to try generating initial manifest again")
-					manifestGenerated, err = manifest.GenerateInitialManifest(appName, appPath, appDirPath)
+					manifestGenerated, err = manifest.GenerateInitialManifest(appData)
 					log.Info("manifestGenerated after generating initial manifest again: ", manifestGenerated)
 					if err != nil {
 						log.Errorf("Error in generating initial manifest: %s", err.Error())
@@ -438,8 +244,8 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 				}
 
 			}
-			log.Infof("[INFO]: Argocd Interlace generates manifest %s", appName)
-			manifestGenerated, err = manifest.GenerateManifest(appName, appDirPath, yamlBytes)
+			log.Infof("[INFO]: Argocd Interlace generates manifest %s", appData.AppName)
+			manifestGenerated, err = manifest.GenerateManifest(appData, yamlBytes)
 			if err != nil {
 				log.Errorf("Error in generating latest manifest: %s", err.Error())
 				return err
@@ -461,7 +267,7 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 		log.Info("buildFinishedOn:", buildFinishedOn, " loc ", loc)
 
 		if interlaceConfig.AlwaysGenerateProv {
-			if !isHelm {
+			if !appData.IsHelm {
 				err = storageBackend.StoreManifestProvenance(buildStartedOn, buildFinishedOn)
 				if err != nil {
 					log.Errorf("Error in storing manifest provenance: %s", err.Error())
@@ -469,7 +275,7 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 				}
 			}
 		} else {
-			if manifestGenerated && !isHelm {
+			if manifestGenerated && !appData.IsHelm {
 				err = storageBackend.StoreManifestProvenance(buildStartedOn, buildFinishedOn)
 				if err != nil {
 					log.Errorf("Error in storing manifest provenance: %s", err.Error())
@@ -482,10 +288,5 @@ func signManifestAndGenerateProvenance(appName, appPath, appClusterUrl,
 		return fmt.Errorf("Could not find storage backend")
 	}
 
-	return nil
-}
-
-func GenerateProvenance(appName, appPath, appClusterUrl,
-	appSourceRepoUrl, appSourceRevision, appSourceCommitSha, appSourcePreiviousCommitSha string, created bool) error {
 	return nil
 }
